+++ src/main/kotlin/com/github/ekvedaras/classfactoryphpstorm/insideClassFactory/definition/ClassPropertyCompletionProviderForAttributesArrayInClosureOfFactoryDefinition.kt
@@ -28,7 +28,7 @@ class ClassPropertyCompletionProviderForAttributesArrayInClosureOfFactoryDefinit
 
         val function = attributesArray.parentOfType<Function>() ?: return
         if (function.parent.parent.parent !is ArrayHashElement) return
-        if (function.parameters[0].name != (attributesArray.firstPsiChild as Variable).name) return
+        if (function.parameters.isEmpty() || function.parameters[0].name != (attributesArray.firstPsiChild as Variable).name) return
 
         val arrayHashElement = function.parent.parent.parent