+++ src/main/kotlin/com/github/ekvedaras/classfactoryphpstorm/insideClassFactory/state/ClassFactoryPropertyReferenceProviderForAttributesArrayKeysInState.kt
@@ -24,7 +24,7 @@ class ClassFactoryPropertyReferenceProviderForAttributesArrayKeysInState : PsiRe
         if (attributesArray.firstPsiChild !is Variable) return PsiReference.EMPTY_ARRAY
 
         val function = attributesArray.parentOfType<Function>() ?: return PsiReference.EMPTY_ARRAY
-        if (function.parameters[0].name != (attributesArray.firstPsiChild as Variable).name) return PsiReference.EMPTY_ARRAY
+        if (function.parameters.isEmpty() || function.parameters[0].name != (attributesArray.firstPsiChild as Variable).name) return PsiReference.EMPTY_ARRAY
 
         if (function.parent.parent.parent !is ArrayHashElement && function.parent.parent.parent !is MethodReference) return PsiReference.EMPTY_ARRAY