+++ src/test/testData/definition/essential/caretAtArrayKeyInClosureWithoutParameters.php
@@ -1,0 +1,24 @@
+<?php
+
+class Account {
+    public function __construct(
+        public readonly string $id,
+        public readonly int $age,
+    ) {}
+}
+
+class AccountFactory extends ClassFactory {
+    protected string $class = Account::class;
+
+    protected function definition(): array
+    {
+        return [
+            'id' => 'abc',
+            'age' => function () {
+                $array = [];
+
+                return $array['<caret>'];
+            }
+        ];
+    }
+}