+++ src/test/testData/make/essential/caretAtArrayKeyMethodCallInShortClosureWithoutParameters.php
@@ -1,0 +1,28 @@
+<?php
+
+class Id {
+    public string $value;
+}
+
+class Account {
+    public function __construct(
+        public readonly string $id,
+        public readonly int $age,
+    ) {}
+}
+
+class AccountFactory extends ClassFactory {
+    protected string $class = Account::class;
+
+    protected function definition(): array
+    {
+        return [
+            'id' => new Id(),
+            'age' => 1,
+        ];
+    }
+}
+
+$attributes = [];
+
+AccountFactory::new()->make(['age' => fn () => $attributes['id']-><caret>]);