+++ src/test/testData/state/outsideFactory/essential/caretAtArrayKeyInClosureWithoutParameters.php
@@ -1,0 +1,18 @@
+<?php
+
+class Account {
+    public function __construct(
+        public readonly string $id,
+        public readonly int $age,
+    ) {}
+}
+
+class AccountFactory extends ClassFactory {
+    protected string $class = Account::class;
+}
+
+AccountFactory::new()->state(['id' => function () {
+    $array = [];
+
+    return $array['<caret>'];
+}])